refactor(navigation): refactor navigation
create connector, screen, navigators
