+++ src/processes/routing/index.ts
@@ -1,0 +1,1 @@
+export { MainTabNavigator } from "./navigators"
