+++ src/processes/routing/navigators/index.ts
@@ -1,0 +1,1 @@
+export { MainTabNavigator } from "./main-tab"
