+++ src/processes/routing/navigators/main-tab/index.ts
@@ -1,0 +1,1 @@
+export { MainTabNavigator } from "./main-tab-navigator"
