+++ src/processes/routing/navigators/main-tab/payments-navigator/index.ts
@@ -1,0 +1,1 @@
+export { PaymentsNavigator } from "./payments-navigator"
