+++ src/processes/routing/screens/index.ts
@@ -1,0 +1,1 @@
+export { PaymentsScreen } from "./payments-screen"
