+++ src/processes/routing/screens/payments-screen.tsx
@@ -1,0 +1,8 @@
+import { PaymentsConnector } from "../../../screens/payments"
+import { PaymentsScreenProps } from "../types"
+
+export const PaymentsScreen = ({ navigation, route }: PaymentsScreenProps) => {
+    return (
+        <PaymentsConnector></PaymentsConnector>
+    )
+}
