+++ src/screens/payments/ui/index.ts
@@ -1,1 +1,2 @@
 export { Payments } from "./payments"
+export { PaymentsConnector } from "./payments-connector"
