+++ src/screens/payments/ui/payments.tsx
@@ -1,3 +1,4 @@
+import { PaymentCategoryUI } from "@shared/api/payment-categories"
 import { Typography } from "@shared/ui/atoms"
 import styled from "styled-components/native"
 
@@ -8,7 +9,11 @@ const Wrapper = styled.View`
     align-items: center;
 `
 
-export const Payments = () => {    
+type Props = {
+    paymentCategories: PaymentCategoryUI[]
+}
+
+export const Payments = ({ paymentCategories }: Props) => {    
     return (
         <Wrapper>
             <Typography align="center">Платежи</Typography>
