+++ src/shared/api/payment-categories/index.ts
@@ -1,8 +1,10 @@
 export { getPaymentCategories } from "./model"
 export { 
-    type PaymentCategoriesAPI, 
+    type PaymentCategoriesAPI,
+    type PaymentCategoryAPI,
     type PaymentServiceAPI,
     type PaymentCategoriesUI,
+    type PaymentCategoryUI,
     type PaymentServiceUI,
     mapPaymentCategoriesToUI
 } from "./types"
